requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
chromadb==0.4.22
together==0.2.11
statsmodels==0.14.1
pillow==10.2.0 